Misc fixes to canonical paths (#868)
This PR fixes a few issues around canonical paths. I started this work
in response to Jacob's request [here](
https://ideaflowteam.slack.com/archives/C07GW72L9RU/p1738129009932409)
but took a broad approach at investigating and fixing problems related
to canonical paths.

Changes:

1. Fixes issue where we'd show the wrong canonical path.

If the canonical relation wasn't available to the user (e.g. the object
is public but the relation is private), the app would assign null as the
canonical relation. But then the next relation we create/load which
points at the object gets automatically assigned as the canonical one.
Even though the object already has a canonical relation (from the dbs
perspective).

The fix is to store the canonical relation id and make the canonical
relation a computed value. When deciding whether to assign a relation as
canonical or not, we only do so if the canonical relation _id_ is null,
as opposed to checking if the _canonical relation_ is falsy (which can
be true when the object does have a canonical relation, it's just that
your client doesn't have access to it)

2. Show "..." prefix on canonical path previews in command bar

With the change above, we can tell whether we actually have the full or
partial canonical path. With partial loading, this is a common state to
be in. Like in the command bar we fetch nodes and relations along
canonical paths but we have a limit to keep the request time reasonably
short. Another change this PR introduces is a new prop on the
`getCanonicalPath` result which tells you whether we've hit the end of
the path or not. If we haven't, we show a "..." as a prefix on the
rendered path.

3. Await canonical path after selecting object from command bar

It's often the case in the command bar that we don't have the full
canonical path available. That means that when you select an option, we
set that partial path on the main view, instead of the full canonical
path. The change in (1) allows us to know when we only have a partial
path, and now we await getting the full path on selection before routing
to it.

This does introduce a bit of a lag, and tbh feels a bit hacky. Ideally
I'd want to route right away, and then have the canonical path load
asynchronously later. My first attempt at doing this worked, but it
caused the tree to re-render in a way that messed with the editing
experience (maybe fixable?). Other attempts were just taking too much
fighting with the tree. This change was the smallest one I could think
of which ensures we render the full canonical path. I'm open to better
ideas though.

Video talking through (2) and (3)
https://www.loom.com/share/bb713f912f7041fcafd59c4f79518e54