Add card view (#837)
Implementation of the card view. The card view is similar to Ideapad.
Each related item is displayed as a card. The cards can have a status
(not acknowledged, acknowledged, in progress, done), as well as comments
and likes - this data is stored in the item's relations with special
labels (such as `status`).