ENT-4625: Creating hashtags behavior (#887)
## Summary

So the plugin implements the new hashtag behaviour. It tracks 2 things:

1. If user just type '#', it creates a new node under the graph hashtag
parent, and replace the text with the mention to that parent.
2. If a user type any normal character (nonspace, alphanumeric, not
meta/shift,...) then we check if the immediate previous node is a
hashtag mention node. If it is then we update it and force editor to
refresh the rendering.

Completes [ENT-4625: Creating hashtags
behavior](https://linear.app/ideaflow/issue/ENT-4625/creating-hashtags-behavior)

- [ ] # should be disabled if the previous key is @ or we're in a
mention state
- [ ] I still have to do the logic of joining them same hashtag together
so we're not creatign a new hashtag with the same name everytime

## How to test


https://github.com/user-attachments/assets/4b8944d7-9481-45d6-8286-cd91a8468482

## Checklist

- [x] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).