fix bug with notifications api (#898)
The fetch to the notifications api assumes the result will always be an
array. But if the auth token is invalid, it handles and returns an
object. Pretty sure this is what's causing some of crashes linked on
slack.

A thing to consider to help avoid stuff like this is improving type
safety across the api. For example
https://trpc.io/docs/client/nextjs/setup