fix: app state to editor wrestles the focus (#929)
## Summary

When we have focused node A and focused node A' (in the sidebar) and we
update A, it automatically update the content and selection in A' and
change the focus over to A'.

The editor doesn't understand that there's 2 DOM editors pointing to the
same tree node, so we have to check that it's the focused DOM element
before we toggle focus.

I'm wary that I might have broken some stuffs from this big PR, but
QA'ing the checklist in this PR seems to be all working
https://github.com/IdeaFlowCo/mew/pull/416

## How to test


https://github.com/user-attachments/assets/fcc9ff7d-1acb-4db6-9c43-ce04210a612e

## Checklist

- [x] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).