chore: fix sublist crashing (#967)
## Summary

I'm not quite sure why `relationsWithPositions` (which is
`allRelationsList`) have relations that are not relating to the current
node.

Only in `Global Root` and `User Root` actually. In the `User Root`
example, the iteration on `nodes.ts` line `712`


```
for (const { relation, position } of treeNode.object.relationsWithPositions.sort((a, b) =>
  comparePositions(a.position, b.position),
)) { 
```

i log it out so that the treeNode of user root has a relation of `user
stream -> random node` which breaks `const childObject =
getOtherObjectOrThrow(relation, treeNode.object.id);`


## How to test

Provide steps on how to test the fix/feature (which you know, because
you tested it 👀)

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [ ] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).