Removed bg color for hashtag counts and aligned the pinned tags divider better (#999)
…tter in sidebar

## Summary

In the sidebar, 
- Removed the background color for hashtag counts.
- Left-aligned the pinned hashtags divider with the hashtag list, and
added top and bottom margins (4px) to it.

Completes
[ENT-4695](https://linear.app/ideaflow/issue/ENT-4695/hashtag-count-styling)

## How to test (AFAIK)

Create hashtags.
Open the sidebar and check if the hashtag counts have no background
color set.
Pin some hashtags in the sidebar and check if the divider is left
aligned with the list of hashtags.

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).