Short URL UX Improvements (#1029)
## Summary

Added toast notifications for saving and removing the short URL.
Changed text in the toast for copying the URL.
Pressing Enter/Return will save changes in the Short URL input field.
Discard unsaved URL changes when display menu is closed.

## Changes
In **`ControlsBar.tsx`**,
- Added `addToast` functions to notify when the URL is saved and removed
(when the `slug.length` is 0).
- Changed toast text for copying the URL (in line 136).
- Added a new `handleKeyDown` function that checks if the pressed key is
Enter and calls the `saveSlug` function when it is.
- Added an `onOpenChange` event handler to the `Popover` component that
detects when the popover is being closed (when `open` becomes `false`).
- When the popover closes, it resets the `slug` state to the `savedSlug`
value using `setSlug(savedSlug)`.

---
Completes
[ENT-4704](https://linear.app/ideaflow/issue/ENT-4704/toast-for-saving-short-url)