Dropdown UI/UX tweaks (#1048)
## Summary

Some minute dropdown tweaks:
- when the mouse is hovered over an option, the option is automatically
selected when the dropdown opens. This isn't optimal because our enter
behaviour assumes nothing is selected by default. This PR makes it so
that the hovered option is only selected when the mouse is moved OR down
arrow is pressed.

## How to test

Provide steps on how to test the fix/feature (which you know, because
you tested it 👀)

## Checklist

- [x] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).