Added bottom margin to Load more button (#1047)
## Summary
This fixes the bottom margin for the Load more button so there is some
space from the node below.

In  **`ChildGroups.tsx`**,
- Added 20px bottom margin to Load more button in line 213.

## How to test
Create a few non-empty nodes. 
In a node above another node, create enough children nodes for the Load
more button to show up.
Check if the button has an appropriate margin from the node below. 

---
Completes
[ENT-4712](https://linear.app/ideaflow/issue/ENT-4712/padding-below-load-more-button)