Fixed: Shift-clicking within the same node selects the whole node (#1103)
## Summary

This fixes an issue where shift-clicking within the current node selects the whole node.

Completes
[ENT-4744](https://linear.app/ideaflow/issue/ENT-4744/shift-clicking-within-the-same-node-should-not-select-the-whole-node)

## Changes 

The issue was created by code added in this PR - [Add shift-click feature
to select multiple nodes](https://github.com/IdeaFlowCo/mew/pull/1080) -
which allows shift-clicking on nodes to select nodes in between the
current node and the clicked node. Because of how this was implemented,
when one shift-clicks on the current node itself, it selects the whole
node instead of selecting the contents within the node.

To fix this, in **`RelatedNodeView.tsx`**,
- Added checks on lines 99 and 124 to make sure the
`handleShiftClickSelection` method is not called if one shift-clicks on
the current node itself.

## How to test

- Click to focus on a node. 
- Shift-click within the same node.
- This should select node contents normally and should not select the
whole node.
- Shift-click on a different node. 
- This should select all the nodes between the focused node and the
shift-clicked node.

## Checklist

- [x] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).