fix: eslint warnings (#1131)
- Fixes most of the eslint warnings (6 left in total across 4 files).
- Added `"@next/next/no-img-element": "off"` to disable `<Image/>`
optimization warnings.
- In certain cases, dependencies such as `refreshTrigger`,
`graphStore.totalNodes`, `customRelation.to` and more, are required to
run the effects again even if not used directly inside the `useEffect`.
- Disabled warnings for with `// eslint-disable-next-line
react-hooks/exhaustive-deps`.
- Removed and added additional dependencies as necessary.
- Note: This update does not introduce `useCallback` changes, so
`useEffect` hooks with missing function dependencies remain unchanged.
Adding missing function dependencies would cause those functions to be
redefined on each render, which in turn would create a rendering loops
as `useEffect` runs again. So we would need to wrap them `useCallback`
and then add their dependencies and so on which is a larger and riskier
change better suited for a separate PR after we test this PR.