convert mention chip to text chip on relation deletion (#1123)
## Summary

When hidden relations turned on, we want the mention chip to turn into a
text chip when the relation is deleted since the mention node will be
disconnected and won't point anywhere when the relation is deleted.


https://github.com/user-attachments/assets/eeff5c05-a4a4-4dd2-ae73-d5f470206604




## How to test

Create a mention chip, with hidden relations enabled, delete the "has
hashtag" part of the node

## Checklist

- [x] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [x] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).