Search Cache Evictions (#1133)
## Summary

Currently, our search results really bloat the local graph store. We can
fix this by making sure that these nodes and edges are evicted after the
search is complete, or after whatever drop-down was open is closed. This
becomes especially clear when you realize that oftentimes our local
graph that we load and is useful to us is actually only a thousand to
two thousand nodes and edges, but once we start searching or use the
search and replace drop-down, the size of our graph store blows up to
30-40 thousand nodes and edges.

@Hackinet would appreciate a second pair of eyes on this.
I've got quite a bit of touching up to do (e.g., removing console log
statements and making the code less decoupled), but the general idea is
here.

## How to test

Run this locally and use the search and replace drop-down a few times.
Watch the console outputs to see how many nodes and edges are being
pulled in from these searches, even after the actual drop-down is
closed.

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).