Stop re-rendering tree on every keystroke (#1157)
### Issue

(copied from Slack
https://ideaflowteam.slack.com/archives/CVD0M5R7F/p1747234096209109?thread_ts=1747157722.366949&cid=CVD0M5R7F):

This is plain wrong. Mew is slow when we type because the tree
re-renders one each keypress, it has to do with how every node inside
Graph is an mobx observable.

Building on that fact, we have a tree getter called get state() that
depends on the root of the tree and that it's children. So when you
update the content in a tree, the state is recomputed, which means a new
root is computed even though only the content of the child node has
changed.

A simple approach would be to just decouple the root and the children
from the state. But it's used in a lot of places, and we do want to
recompute state in some cases where, say when we do bulk node operations
as moving nodes or deleting them.

So we do need a way to not trigger a re-render when a node content is
updated. Maybe we can access the nodes through a proxy, to have
selective observability. Or use custom mobx reactions (with selective
observation properties) but then we need to have a
ReactionDisposeManager of sorts, that runs the disposers of stale trees
and that's laggy too.

The memory spike happens because when we type, a new tree(s) is
re-rendered. The old tree hasn't been garbage collected from the heap
memory. If you run the performance tools, you would see frequent minor
GCs and sometimes a big GC. Frequent minor GCs (pauses the thread) and
the tree re-rendering is what makes it laggy. I have previously checked
all event listeners, they all have a cleanup and all are accounted for.
(edite

### Fix:
Funnily enough, it was just plain old binary search, after trying
different approaches, creating object proxies, proxy getters,
memoization, dirty state tracking, manual state computation, reaction
disposers and more, all of them ended up having some nasty side effects.

The problem was that the tree state was a computed property, in the
hydration, we wrapped the graph object inside tree nodes. We knew the
problem was because the content changed, so the chain: state ->
hydration method -> nodes/groups would recompute.

We were trying to not recompute the tree state when content changed but
also keep the recompilation reactive when the relations changed, or the
positions changed. While writing a custom reaction in the tree and it's
evaluation expression, I realised I wasn't tracking the objects really
since I was trying to be optimal about it.

My code was, if graph's node length change OR relations change OR
position change, recompute, else return the tree state from cache. So I
thought, okay, maybe instead of doing selective content property
untracking, we can just untrack the whole object when we are creating a
tree. Just track the relations and everything else. When nodes are
created/deleted, relations are created/deleted too.

So I wrapped the whole hydration logic in untracked() , only problem was
nothing worked because we stopped observing relations too. I then
divided the function into two sections. Wrapped the hydration function
call in untracked() and the return in untracked(). Went into hydration
login, kept wrapping and unwrapping the whole functions, then parts of
it in untracked() , isolating the parts where mobx establishes an
observer for the objects but not relations/positions.