Mobile screen casework (#1182)
## Summary

Makes the mobile experience more intuitive. If QC is open and the user
is on a mobile (small) screen, the FAB makes a new node in the quick
capture tree root instead of zooming into a new node in the main view.


## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).