ENT-4772: Refactor user flow (#1186)
Changes:

- Stop initialising the app as an unauthenticated user. Load the user
from local storage instead.
- Avoid fetching user and settings data on every load. Only fetch from
the server if the user doesn’t exist or the JWT has expired.
- Move the user loading logic to `useSetupUser` hook to clean up
StoreProvider.tsx.
- Remove slug pre-fetching and `objectId` computation on the client
side. Load the initial `objectId` on the server based on the requested
URL. Since NextJS doesn’t allow direct access to the requested URL, use
a middleware instead.
- See comment and replies:
https://github.com/vercel/next.js/issues/43704#issuecomment-2090798307
- Stop passing the authedFetch in the `GraphStore` (larger plan to
eventually remove it in favour of `axios` but the PR was getting a bit
long).