A missing case for mlnote destruction (#1217)
## Summary

When a multiline note is like so:
```
a
b<<cursor
```
and we press tab, the MLNote should be destroyed. It's not being
destroyed right now.

## How to test

Try it out.

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Bug Fixes**
- Improved handling of indenting the second child in multiline notes to
preserve note structure and prevent unintended nesting.
- Selection focus is now correctly updated after indenting in this edge
case for a smoother editing experience.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->