ENT-4896: Fix broken paste when copied from inside the app (#1220)
When text is copied inside from the app, the clipboard looks like this
with type `text/html`:
```
<span style="white-space: pre-wrap;">Text</span>
```

The problem happens when`<span>` is not wrapped inside a `<div>` or
`<p>`.
The `walk` method doesn't have any `span` case inside the switch
directly, it assumes that a `span` would always come wrapped in a `<p>`.
So `span` fallbacks to default case. In the default case, we determine
if the `<span>` has children.

The code assumes a span could have another `<span>` or `<b>` as `<i>`
(`ElementNode`s). If so, recurse into `walk` method again with a new
depth with children, if not, do nothing.

The thing is text in browsers are wrapped in `TextNode`, so a `span`
will _always_ have children. We were recursing into `walk` but we were
throwing away nodes with type `TextNode` with:

```
if (node.nodeType !== Node.ELEMENT_NODE){
        continue;
}
```

The fix uses/pushes text from a `TextNode` (leaf node) when encountered
instead of throwing it away.

Testing:
- You can use the inspector here to see clipboard:
https://evercoder.github.io/clipboard-inspector/