Add a Root Descendants Node filter for the Updates Feed (#1228)
## Summary

Please include a summary of the changes, and how it addresses issue or
feature. Explain changes to dependencies (with link to its homepage) and
environment variables.
- Created a new function in Tree.ts: getDescendantNodes(). This will go
down the current root node, and return all the descendant nodes.
- Created new function in ViewStore.ts, which will get the result from
treeView.getDescendantNodes and map them to a set by object ID.
- Refactored UpdatesFeed.tsx such that the "updates" array uses a state.
This makes it so that it can be dynamically updated, and triggers a
refresh for the UI.
- Added a new checkbox in UpdatesFeed to enable/disable filtering by
Root's Descendants
- Added a filter function in UpdatesFeed that will check if the current
list of updates' relevant node IDs (depending on the type of update) is
a member of the current root's descendants.

Completes ENT-4806 (or one of the other [magic
words](https://linear.app/docs/github#magic-words) to link the PR to a
linear issue)

## How to test
1. Start Mew
2. Create two Child Nodes from the Global Root
3. Create a GrandChild to the first child node
4. From Global Root, Go to Updates Feed. Check the "Filter Updates by
Root" box. It should do nothing.
5. Go back to Your Root, enter the first child as root, return to
Updates Feed
6. Check the "Filter Updates by Root" box, it should now only show the
changes for this current root and the grandchild node.

Provide steps on how to test the fix/feature (which you know, because
you tested it 👀)

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X ] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Added a filter option to the Updates Feed to show updates related only
to descendant nodes.
- Introduced a checkbox in the Updates Feed for toggling the filter,
improving update navigation.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->