ENT-4817: Introduce canonical path caching and auto update (#1234)
Changes:
- Add new table for canonical path caching.
- Add a mobx store class `CanonicalPathCacheStore`.
- During comouting canonical path, check if we have reached an end state
for an objectId, if so:
- Check if a cache exists on the server, and we have downloaded that
cache, compare it with the local path. Since we reached en end state
locally, it's guaranteed a local path exists. If there's a mismatch in
length or content, update the cache locally and on server.
- During rendering a a canonical path, check if we haven't reached an
end state and if a cached path exists, if so, render it.

There's a few future scopes/improvements here:
https://linear.app/ideaflow/issue/ENT-4817/improve-canonical-paths-and-caching