Refactor some load functions in loadGraphData (#1241)
## Summary
Found that some of the functions that are calling fetchAndLoad aren't
written correctly.

1. loadRelationTypes is unused.
2. loadWithText is saving the return value of fetchAndLoad and not using
it.
3. loadWithCanonicalIds is returning the value of fetchAndLoad, but none
of the callers are using it.

This is a fairly small but non-functional cleanup.

## How to test
I think the best way is to just look at the code, since none of what I
did should affect anything. localRelationTypes has no uses. nodeIds is
not being used anywhere, loadWithCanonicalIds is not being called
anywhere for its return statement.

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [ ] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).