Redefine isExpanded to only allow for tree nodes with children (#1251)
## Summary
As illustrated by Samuel's example in:
https://www.loom.com/share/57070cd64c0b43eda8b1566a425ce4d0

What happens is that when you delete the last child node of a parent, it
doesn't actually reset the parent's status as collapsed.

Then, when you delete the parent node, when it gets merged to the above
node, it's status as expanded gets propagated up to the node above.
You can see in the loom video at 1:16, the parent node's arrow is still
point down, so the UI is treating it as if it's expanded.

I tried creating a callback in the delete functionality, to set the
parent's expanded status to be false if there are no more children. But
was having some trouble with accessing the TreeNode from the GraphStore
(which is touching GraphNodes).

For now this fix is a good redundancy. It also fixes existing states
where if you deleted the children in a previous session, it'll still
show it as expanded. Example:

![image](https://github.com/user-attachments/assets/36773833-e0bf-4b48-adff-49bc9bf61ee6)
Dinosaur has no children, yet the arrow is facing down.

This fix should change it so that it'll point right, even if the status
is "Expanded".

Completes
[ENT-4831](https://linear.app/ideaflow/issue/ENT-4831/backspacing-node-into-sibling-above-sometimes-expands-its-children)

## How to test
Create 2 pairs of parent child pairs:
A
-- B
C
-- D

Collapse A to hide B.

Delete D
C should now be collapsed (rightward arrow)
When you delete C, A should not expand.

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [ ] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).