Prevent rubberbanding for non-editable nodes. Don't allow merging from below nodes. (#1256)
## Summary
Addresses the 2nd case in this video:
https://www.loom.com/share/6165ef9ec5944cb1827dd9c1c95d6d5e

This issue is caused when we allow backspace triggered node-merging when
the nextNodeAbove is edit restricted. The request goes through but get's
denied, so we get rubberbanding behavior.

Fixed this by checking the nextNodeAbove's isEditRestricted value before
going through with the mergeNode.

I also caught two other cases with edit restricted nodes, where you can
edit the prefix and suffix of a node, and it rubberbands.

## How to test
Create a new node, then backspace it into the pinned section.
It should not allow you to backspace now (won't see any rubberbanding).

In the pinned section for the "My Favorites" node, try typing at the
start and end of the input. It should not allow any text inputs now. (no
rubberbanding).

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).