Fix bug with editor selection not storing when you pin the node (#1254)
## Summary
https://www.loom.com/share/ec186969f9ba4fc79dd3ecd41e1ebe6b

When you have an editor selection for a node, and then pin the node, you
lose the editor selection.

I added a callback so that we keep a reference to the previous node,
find the input element via an html selector, then re-set the focus to
that after the DOM updates.

## How to test
Disable the setting: "Hide pinned elements from the main list"

![image](https://github.com/user-attachments/assets/4fbb45ae-cab6-4a26-9aee-a9ffb187116e)

Select a node such that your cursor is editing it's input.
Pin the item
Your input should still be selected

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).