Trim Content on search results (#1266)
## Summary

Trim the content on search results if it's over a threshold of chars.

## How to test

Open up the S&R dropdown or the command bar dropdown and search for
something that has a lot of text.

Next steps: apply to relations as well

## Checklist

- [X] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [X] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).