Fix bug about collapsing parent (#1258)
## Summary
https://www.loom.com/share/16c342428c63438396c65b3d21d90801

This was caused by my change in
https://github.com/IdeaFlowCo/mew/commit/69d7eca5c324dbca3c409881754793135394b687.

What we did was to collapse/expand the target node to whatever state the
source was. However, if the target was expanded already, and the source
is collapsed, then we don't want to collapse the target.

This happens in this configuration:
target
----source <-
----anotherChild

If you delete source, which is collapsed, it will collapse target.

## How to test
Reproduce steps from loom video.

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).