Merge weirdness (#1261)
## Summary
Addresses the first bug of this video
https://www.loom.com/share/6165ef9ec5944cb1827dd9c1c95d6d5e

It turns out all of the existing logic was mostly working, there just
was a line that was setting the index to the end of the input, even if
the set position wasn't explicitly set to "end".

Also test the creation of a new node:
https://www.loom.com/share/cb6af3d8126641808c56670d427596d7

## How to test
Create a node and a node below it (not a child)
Hit backspace from the bottom node to merge this node with the above
node
The index of the selector should be at the middle of the 2 strings

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).

Co-authored-by: Samuel Xu <samuelzxu@gmail.com>