Make "My Streams" EditRestricted and DeleteRestricted (#1276)
## Summary
As discussed with Hari, Samuel, and Piyush, we found that "My Streams"
should behave the same way as "My Hashtags", "My Favorites", and "My
Templates", as in they shouldn't be editable, and deletable.

## How to test
Try to edit the node and delete it, it should behave the same as the
others now.

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [ ] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).